Refactor project structure and update TypeScript configurations
- Added compatibility types for Next.js navigation in `next-env.d.ts`.
- Updated Tailwind CSS configuration to include the new App Router directory.
- Modified TypeScript configuration to include `.next/types/**/*.ts` in the include path.
- Removed deprecated `_app.tsx`, `dashboard.tsx`, `index.tsx`, and API routes related to Privy authentication.
- Cleaned up unused imports and components to streamline the codebase.
