+++ app/layout.tsx
@@ -1,0 +1,65 @@
+import "../styles/globals.css";
+import { PrivyProvider } from "@privy-io/react-auth";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Privy Auth Starter",
+  description: "Privy Auth Starter",
+  icons: {
+    icon: [
+      { url: "/favicons/favicon.ico", sizes: "any" },
+      { url: "/favicons/icon.svg", type: "image/svg+xml" },
+    ],
+    apple: "/favicons/apple-touch-icon.png",
+  },
+  manifest: "/favicons/manifest.json",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <head>
+        <link
+          rel="preload"
+          href="/fonts/AdelleSans-Regular.woff"
+          as="font"
+          crossOrigin=""
+        />
+        <link
+          rel="preload"
+          href="/fonts/AdelleSans-Regular.woff2"
+          as="font"
+          crossOrigin=""
+        />
+        <link
+          rel="preload"
+          href="/fonts/AdelleSans-Semibold.woff"
+          as="font"
+          crossOrigin=""
+        />
+        <link
+          rel="preload"
+          href="/fonts/AdelleSans-Semibold.woff2"
+          as="font"
+          crossOrigin=""
+        />
+      </head>
+      <body>
+        <PrivyProvider
+          appId={process.env.NEXT_PUBLIC_PRIVY_APP_ID || ""}
+          config={{
+            embeddedWallets: {
+              createOnLogin: "all-users",
+            },
+          }}
+        >
+          {children}
+        </PrivyProvider>
+      </body>
+    </html>
+  );
+}
