+++ tailwind.config.js
@@ -2,7 +2,10 @@ const defaultTheme = require('tailwindcss/defaultTheme');
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
+  content: [
+    './app/**/*.{js,ts,jsx,tsx,mdx}', // App Router directory
+    './components/**/*.{js,ts,jsx,tsx}' // Components directory
+  ],
   theme: {
     extend: {
       fontFamily: {
